requirements.txt
git+https://github.com/Rapptz/discord.py.git@b561024
Pillow
numpy
pandas
lxml
requests
bs4
nest_asyncio
pyautogui
websocket-client
python-dotenv
pyjwt
html5lib
asgiref==3.5.2
Django==4.0.6
sqlparse==0.4.2
tzdata==2022.1
pygments
psycopg2